InfoRes code moved out of Source.py into its own utility file
