+++ .gitignore
@@ -5,9 +5,7 @@ __pycache__
 *.egg-info
 tests/logs/*.log
 build
-target
 dist
-docker_test_*
 
 /build
 /dist
@@ -17,9 +15,9 @@ venv
 .idea/
 
 # test directories
+tests/logs
 docker_test_*
 target
-tests/logs
 
 # the stray output file from a test-merge.yaml driven test
 merged-kg_stats.yaml
