+++ kgx/source/source.py
@@ -1,7 +1,6 @@
-from typing import Dict, Generator, Any, Union, Optional, Tuple, List, Set, Callable
-import re
+from typing import Dict, Union, Optional
 
-from kgx.utils.kgx_utils import knowledge_provenance_properties, column_types
+from kgx.utils.infores import InfoResContext
 from kgx.prefix_manager import PrefixManager
 from kgx.config import get_logger
 
@@ -21,8 +20,7 @@ class Source(object):
         self.node_properties = set()
         self.edge_properties = set()
         self.prefix_manager = PrefixManager()
-        self.default_provenance = 'Graph'
-        self._infores_catalog: Dict[str, str] = dict()
+        self.infores: Optional[InfoResContext] = None
 
     def set_prefix_map(self, m: Dict) -> None:
         """
@@ -35,422 +33,6 @@ class Source(object):
 
         """
         self.prefix_manager.update_prefix_map(m)
-
-    def _infores_processor(
-            self,
-            ksf: str,
-            infores_rewrite_filter: Optional[Tuple] = None
-    ) -> Callable:
-        """
-        Full processor of a Knowledge Source name into an InfoRes. The conversion is made based on
-        client-caller specfied rewrite rules for a given knowledge source field ('ksf').
-
-        Parameters
-        ----------
-        ksf: str
-            Knowledge source field being processed.
-
-        infores_rewrite_filter: Optional[Tuple]
-            The presence of this optional Tuple argument signals an InfoRes rewrite of any
-            Biolink 2.0 compliant knowledge source field name in node and edge data records.
-            The mere presence of a (possibly empty) Tuple signals a rewrite. If the Tuple is empty,
-            then only a standard transformation of the field value is performed. If the Tuple has
-            an infores_rewrite[0] value, it is assumed to be a regular expression (string) to match
-            against. If there is no infores_rewrite[1] value or it is empty, then matches of the
-            infores_rewrite[0] are simply deleted from the field value prior to coercing the field
-            value into an InfoRes CURIE. Otherwise, a non-empty second string value of infores_rewrite[1]
-            is a substitution string for the regex value matched in the field. If the Tuple contains
-            a third non-empty string (as infores_rewrite[2]), then the given string is added as a prefix
-            to the InfoRes.  Whatever the transformations, unique InfoRes identifiers once generated,
-            are used in the meta_knowledge_graph and also reported using the get_infores_catalog() method.
-
-        Returns
-        -------
-        Callable
-            A locally configured Callable that knows how to process
-            a source name string into an infores CURIE, using on client-specified
-            rewrite rules applied alongside standard formatting rules.
-
-        """
-
-        # Check for non-empty infores_rewrite_filter
-        if infores_rewrite_filter:
-            _filter = re.compile(infores_rewrite_filter[0]) if infores_rewrite_filter[0] else None
-            _substr = infores_rewrite_filter[1] if len(infores_rewrite_filter) > 1 else ''
-            _prefix = infores_rewrite_filter[2] if len(infores_rewrite_filter) > 2 else ''
-        else:
-            _filter = None
-            _substr = ''
-            _prefix = ''
-
-        def _get_infores(source: str) -> str:
-            """
-            Get InfoRes CURIE inferred from source name.
-
-            Parameters
-            ----------
-            source: str
-                Name of Information Resource associated with the InfoRes
-                (i.e. from which the InfoRes was inferred)
-
-            Returns
-            -------
-            str:
-                infores CURIE, retrieved or generated.
-
-            """
-            if source in self._infores_catalog:
-                return self._infores_catalog[source]
-            else:
-                infores: str = _process_infores(source)
-                if infores:
-                    self._infores_catalog[source] = infores
-                    return infores
-                else:
-                    return ''
-
-        def _process_infores(source: str) -> str:
-            """
-            Process a single knowledge Source name  string into an infores, by applying rules
-            in the _infores_processor() closure context, followed by standard formatting.
-
-            Parameters
-            ----------
-            source: str
-                Knowledge source name string being processed.
-
-            Returns
-            -------
-            str
-                Infores CURIE inferred from the input Knowledge Source name string.
-
-            """
-            # don't touch something that already looks like an infores CURIE
-            if source.startswith('infores:'):
-                return source
-
-            if _filter:
-                infores = _filter.sub(_substr, source)
-            else:
-                infores = source
-            infores = _prefix + ' ' + infores
-            infores = infores.strip()
-            infores = infores.lower()
-            infores = re.sub(r"\s+", "_", infores)
-            infores = re.sub(r"\.+", "_", infores)
-            infores = re.sub(r"[\W]", "", infores)
-            infores = re.sub(r"_", "-", infores)
-
-            # TODO: to be fully compliant, the InfoRes needs to have the 'infores' prefix?
-            infores = "infores:"+infores
-
-            return infores
-
-        def parser_list(sources: Optional[List[str]] = None) -> List[str]:
-            """
-            Infores parser for a list of input knowledge source names.
-
-            Parameters
-            ----------
-            sources: List[str]
-                List of Knowledge source name strings being processed.
-
-            Returns
-            -------
-            List[str]
-                Source name strings transformed into infores CURIES, using _process_infores().
-
-            """
-            if not sources:
-                return [self.default_provenance]
-            results:  List[str] = list()
-            for source in sources:
-                infores = _get_infores(source)
-                if infores:
-                    results.append(infores)
-            return results
-
-        def parser_scalar(source=None) -> str:
-            """
-            Infores parser for a single knowledge source name string.
-
-            Parameters
-            ----------
-            source: str
-                Knowledge source name string being processed.
-
-            Returns
-            -------
-            str
-                Source name string transformed into an infores CURIE, using _process_infores().
-
-            """
-            return self.default_provenance if not source else _get_infores(source)
-
-        if ksf in column_types and column_types[ksf] == list:
-            return parser_list
-        else:
-            # not sure how safe an assumption for all non-list column_types, but...
-            return parser_scalar
-
-    @staticmethod
-    def _infores_default(ksf, default=None):
-        """
-        Lightweight alternative to _infores_processor which simply assigns knowledge_source fields client-defined
-        simple default knowledge source strings (not constrained to be formatted as infores CURIEs).
-
-        Parameters
-        ----------
-        ksf: str
-            Knowledge source field being processed.
-
-        Returns
-        -------
-        Callable
-            A locally configured Callable that knows how to process a source name string
-            (possibly empty) into a suitable (possibly default) infores string identifier.
-
-        """
-
-        def default_value_list(sources: List[str] = None):
-            """
-            Infores default method for a list of input knowledge source names.
-
-            Parameters
-            ----------
-            sources: List[str]
-                List of Knowledge source name strings being processed.
-
-            Returns
-            -------
-            List[str]
-                Infores identifiers mapped to input source strings.
-
-            """
-            if not default:
-                return list()
-            if not sources:
-                return [default]
-            else:
-                return sources
-
-        def default_value_scalar(source=None):
-            """
-            Infores default method for single input knowledge source name.
-
-            Parameters
-            ----------
-            source: str
-                Knowledge source name string being processed.
-
-            Returns
-            -------
-            str
-                Infores identifier mapped to the input source string.
-
-            """
-            if not default:
-                return None
-            if not source:
-                return default
-            else:
-                return source
-
-        if ksf in column_types and column_types[ksf] == list:
-            return default_value_list
-        else:
-            # not sure how safe an assumption for non-list column_types, but...
-            return default_value_scalar
-
-    def set_provenance_map_entry(self, ksf: str, ksf_value: Any, entry: Dict, key: str):
-        if isinstance(ksf_value, str):
-            ksf_value = ksf_value.strip()
-            if ksf_value.lower() == 'true':
-                entry[key] = self._infores_processor(ksf)
-            elif ksf_value.lower() == 'false':
-                entry[key] = self._infores_default(ksf)  # source suppressed
-            else:
-                entry[key] = self._infores_default(ksf, ksf_value)
-        elif isinstance(ksf_value, bool):
-            if ksf_value:
-                entry[key] = self._infores_processor(ksf)
-            else:  # false, ignore this source?
-                entry[key] = self._infores_default(ksf)  # source suppressed
-        elif isinstance(ksf_value, (list, set, tuple)):
-            entry[key] = self._infores_processor(ksf, infores_rewrite_filter=ksf_value)
-
-    def set_provenance_map(self, kwargs: Dict):
-        """
-        A knowledge_source property indexed map set up with various graph_metadata
-        Callable methods to process input knowledge source values into
-        suitable InfoRes identifiers.
-
-        Parameters
-        ----------
-        kwargs: Dict
-            The input keyword argument dictionary was likely propagated from the
-            Transformer.transform() method input_args, and is here harvested for
-            static defaults or rewrite rules for knowledge_source slot InfoRes value processing.
-
-        """
-        if 'default_provenance' in kwargs:
-            self.default_provenance = kwargs.pop('default_provenance')
-
-        # Biolink 2.0 knowledge_source 'knowledge_source' derived fields
-        ksf_found = False
-        for ksf in knowledge_provenance_properties:
-            if ksf in kwargs:
-                if not ksf_found:
-                    ksf_found = ksf  # save the first one found, for later
-                ksf_value = kwargs.pop(ksf)
-                # Check if the ksf_value is a multi-valued catalog of patterns for a
-                # given knowledge graph field, indexed on each distinct regex pattern
-                if isinstance(ksf_value, dict):
-                    for ksf_pattern in ksf_value.keys():
-                        if ksf not in self.graph_metadata:
-                            self.graph_metadata[ksf] = dict()
-                        self.set_provenance_map_entry(
-                            ksf,
-                            ksf_value[ksf_pattern],
-                            self.graph_metadata[ksf],
-                            ksf_pattern
-                        )
-                else:
-                    self.set_provenance_map_entry(
-                        ksf,
-                        ksf_value,
-                        self.graph_metadata,
-                        ksf
-                    )
-
-        # if none specified, add at least one generic 'knowledge_source'
-        if not ksf_found:
-            ksf_found = 'knowledge_source'  # knowledge source field 'ksf' is set, one way or another
-            if 'name' in kwargs:
-                self.graph_metadata['knowledge_source'] = self._infores_default(ksf_found, kwargs['name'])
-            else:
-                self.graph_metadata['knowledge_source'] = self._infores_default(ksf_found, self.default_provenance)
-            
-
-        # TODO: better to lobby the team to totally deprecated this, even for Nodes?
-        if 'provided_by' not in self.graph_metadata:
-            self.graph_metadata['provided_by'] = self.graph_metadata[ksf_found]
-
-    def set_provenance(self, ksf: str, data: Dict):
-        """
-            Compute the knowledge_source value for the current node or edge data, using the
-            infores rewrite context previously established by a call to set_provenance_map().
-
-            Parameters
-            ----------
-            ksf: str
-                Knowledge source field being processed.
-            data: Dict
-                Current node or edge data entry being processed.
-
-        """
-        if ksf not in data.keys():
-            if ksf in self.graph_metadata and not isinstance(self.graph_metadata[ksf], dict):
-                data[ksf] = self.graph_metadata[ksf]()  # get default ksf value?
-            else:
-                # if unknown ksf or is an inapplicable pattern
-                # dictionary, then just set the value to the default
-                data[ksf] = [self.default_provenance]
-        else:  # valid data value but... possible InfoRes rewrite?
-            # If data is s a non-string iterable
-            # then, coerce into a simple list of sources
-            if isinstance(data[ksf], (list, set, tuple)):
-                sources = list(data[ksf])
-            else:
-                # Otherwise, just assumed to be a scalar
-                # data value for this knowledge source field?
-                # Treat differentially depending on column type...
-                if column_types[ksf] == list:
-                    sources = [data[ksf]]
-                else:
-                    sources = data[ksf]
-            if ksf in self.graph_metadata:
-                if isinstance(self.graph_metadata[ksf], dict):
-                    # Need to iterate through a pattern dictionary
-                    for pattern in self.graph_metadata[ksf].keys():
-                        for source in sources:
-                            # TODO: I need to test pattern match to each source?
-                            data[ksf] = self.graph_metadata[ksf][pattern](source)
-                            if data[ksf]:
-                                break
-                else:
-                    data[ksf] = self.graph_metadata[ksf](sources)
-            else:  # leave data intact?
-                data[ksf] = sources
-
-        # ignore if again empty at this point
-        if not data[ksf]:
-            data.pop(ksf)
-
-    def set_node_provenance(self, node_data: Dict):
-        """
-        Sets the node knowledge_source value for the current node. At the moment, nodes are still
-        hard-coded to using the (Biolink 2.0 deprecated) 'provided_by' knowledge_source property;
-        However, this could change in the future to use edge 'knowledge_source' properties.
-
-        Parameters
-        ----------
-        node_data: Dict
-            Current node data entry being processed.
-
-        """
-        self.set_provenance('provided_by', node_data)
-
-    # TODO: need to design a more efficient algorithm here...
-    def set_edge_provenance(self, edge_data: Dict):
-        """
-        Sets the node knowledge_source value for the current node. Edge knowledge_source properties
-        include the full Biolink 2.0 'knowledge_source' related properties.
-
-        Parameters
-        ----------
-        edge_data: Dict
-            Current edge data entry being processed.
-
-        """
-        ksf_found = False
-        data_fields = list(edge_data.keys())
-        for ksf in data_fields:
-            if ksf in knowledge_provenance_properties:
-                ksf_found = True
-                self.set_provenance(ksf, edge_data)
-        if not ksf_found:
-            for ksf in self.graph_metadata:
-                if ksf != 'provided_by':
-                    self.set_provenance(ksf, edge_data)
-
-    def get_infores_catalog(self) -> Dict[str, str]:
-        """
-        Retrieves the catalogs mapping of Knowledge Source names to an InfoRes.
-
-        Returns
-        -------
-        Dict[str, Set[str]]
-            Dictionary where the index string is an InfoRes CURIE, values are sets of Knowledge Source Names
-
-        """
-        return self._infores_catalog
-
-    def parse(self, **kwargs: Any) -> Generator:
-        """
-        This method reads from the underlying store, using the
-        arguments provided in ``config`` and yields records.
-
-        Parameters
-        ----------
-        **kwargs: Any
-
-        Returns
-        -------
-        Generator
-
-        """
-        pass
 
     def check_node_filter(self, node: Dict) -> bool:
         """
@@ -650,4 +232,23 @@ class Source(object):
         generally a Callable function. This operation can be used in the code when the metadata is
         no longer needed, but may cause peculiar Python object persistent problems downstream.
         """
-        self.graph_metadata.clear()
+        self.infores = None
+
+    def set_provenance_map(self, kwargs):
+        """
+        Set up a provenance (Knowledge Source to InfoRes) map
+        """
+        self.infores = InfoResContext(self)
+        self.infores.set_provenance_map(kwargs)
+
+    def set_node_provenance(self, node_data):
+        """
+        Set a specific node provenance value.
+        """
+        self.infores.set_node_provenance(node_data)
+
+    def set_edge_provenance(self, edge_data):
+        """
+        Set a specific edge provenance value.
+        """
+        self.infores.set_edge_provenance(edge_data)
